api/requirements.txt
Flask >=0.12.2, <1
flask-cors >=3.0.3, <4
SQLAlchemy >=1.1.13, <2
Flask-SQLAlchemy >=2.2, <3
psycopg2 >=2.7.3, <3 --no-binary :all:
requests >=2.18.4, <3
boto3 >=1.4.6, <2
xlrd >=1, <2
xlutils >=2, <3
openpyxl >=2.5.0a3, <3
python-dateutil >=2.6.1, <3
Celery >=4.1.0, <5
pycurl >=7.43.0, <8
Flask-SSE >=0.2.1, <1
gevent >=1.2.2, <2

